Combine CameraAnimationsManager source files
